Making migrations work outside of rails.
